+++ controllers/api.js
@@ -9,15 +9,31 @@ function getTweets(req, res) {
         access_token_secret: process.env.ACCESS_TOKEN_SECRET
       });
     var params = {screen_name: 'valeriaoshiro'};
-    client.get('statuses/user_timeline', params, function(error, tweets, response) {
+    
+    var cursor = -1;
+    var api_path = 'friends/list.json?screen_name=twitterapi&skip_status=true&include_user_entities=false&count=200';
+    var celebrities = [];
+    
+    client.get('friends/list.json?cursor=-1&screen_name=twitterapi&skip_status=true&include_user_entities=false&count=200', params, function(error, tweets, response) {
     if (!error) {
-        console.log(tweets);
+        //cursor = tweets.next_cursor;
+        //console.log("////////", tweets.next_cursor, tweets.users.length);
+        //console.log("*********users ", tweets.users);
+        tweets.users.forEach(function(user){
+            if(user.verified) celebrities.push(user.name);
+            console.log("*********", celebrities);
+        })
     } else {
         console.log(error);
     }
     });
 }
 
+function getMovies(req, res) {
+    // hello
+}
+
 module.exports = {
-    getTweets
+    getTweets,
+    getMovies
 }
