+++ routes/routes.js
@@ -3,5 +3,6 @@ var router = express.Router();
 var apiCtrl = require('./../controllers/api');
 
 router.get('/tweets', apiCtrl.getTweets);
+router.get('/movies', apiCtrl.getMovies);
 
 module.exports = router;
